+++ discord/message.py
@@ -785,15 +785,11 @@ class Message(Hashable):
         The call information associated with this message, if applicable.
 
         .. versionadded:: 2.6
-    data: :class:`dict`
-        The data payload that was received to create this message.
-
-        .. versionadded:: 2.7
     """
 
     __slots__ = (
         "_state",
-        "data",
+        "_raw_data",
         "_edited_timestamp",
         "_cs_channel_mentions",
         "_cs_raw_mentions",
@@ -847,7 +843,7 @@ class Message(Hashable):
         data: MessagePayload,
     ):
         self._state: ConnectionState = state
-        self.data: MessagePayload = data
+        self._raw_data: MessagePayload = data
         self.id: int = int(data["id"])
         self.webhook_id: int | None = utils._get_as_snowflake(data, "webhook_id")
         self.reactions: list[Reaction] = [