+++ CHANGELOG.md
@@ -36,7 +36,7 @@ These changes are available on the `master` branch, but have not yet been releas
   `Permissions.use_external_sounds` and
   `Permissions.view_creator_monetization_analytics`.
   ([#2620](https://github.com/Pycord-Development/pycord/pull/2620))
-- Added `Message.data` attribute.
+- Added `Message._raw_data` attribute.
   ([#2670](https://github.com/Pycord-Development/pycord/pull/2670))
 
 ### Fixed